data-transfer/requirements.txt
mysql-connector-python
pymongo
